pyproject.toml
[build-system]
requires = ["setuptools >= 68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finops-agent"
version = "0.1.0"
description = "Open-source, CLI-first, multi-cloud infrastructure cost reasoning agent"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">= 3.11"
authors = [
    { name = "finops-agent contributors" },
]
keywords = ["finops", "cloud", "cost", "aws", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "boto3 >= 1.34",
    "typer >= 0.9",
    "rich >= 13.0",
    "pyyaml >= 6.0",
    "openai >= 1.0",
    "anthropic >= 0.18",
    "httpx >= 0.27",
]

[project.optional-dependencies]
gcp = [
    "google-cloud-bigquery >= 3.11",
    "google-cloud-compute >= 1.14",
    "google-cloud-container >= 2.21",
    "google-auth >= 2.0",
]
azure = [
    "azure-identity >= 1.15",
    "azure-mgmt-costmanagement >= 4.0",
    "azure-mgmt-compute >= 30.0",
    "azure-mgmt-network >= 25.0",
    "azure-mgmt-containerservice >= 28.0",
    "azure-mgmt-resource >= 23.0",
]
dev = [
    "pytest >= 8.0",
    "pytest-cov >= 4.0",
    "ruff >= 0.3",
    "mypy >= 1.8",
    "boto3-stubs[ec2,elbv2,ce,eks,cloudwatch] >= 1.34",
    "types-PyYAML >= 6.0",
]

[tool.setuptools.packages.find]
include = ["cli*", "cloud*", "collectors*", "cost_model*", "intelligence*", "llm*", "storage*", "scheduler*"]

[project.scripts]
finops = "cli.main:app"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501", "SIM117"]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["cloud.aws.*"]
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: requires real cloud credentials (deselect with '-m not integration')",
]
